requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.8
django-config==0.1.4
django-cors-headers==4.9.0
django-filter==25.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
qrcode==8.2
sqlparse==0.5.3
whitenoise==6.11.0
